Everything to float

+++ MoogleEngine/Coisine_Sim.cs
@@ -8,17 +8,17 @@ namespace MoogleEngine
 {
     public static class Coisine_Sim
     {
-        private static double Magnitude(Vector v1)
+        private static float Magnitude(Vector v1)
         {
-            return Math.Sqrt(v1 * v1);
+            return (float)Math.Sqrt(v1 * v1);
         }
-        private static double CoisineSimilarity(Vector general, Vector query)
+        private static float CoisineSimilarity(Vector general, Vector query)
         {
             if ((Magnitude(general) * Magnitude(query)) == 0)
                 return 0;
             return (general * query) / (Magnitude(general) * Magnitude(query));
         }
-        public static Dictionary<int, double> GetCoisineSim(double[] query_array, double[][] documents_matrix)
+        public static Dictionary<int, float> GetCoisineSim(float[] query_array, float[][] documents_matrix)
         {
             List<Vector> document_vectors = new List<Vector>();
             foreach (var array in documents_matrix)
@@ -26,7 +26,7 @@ namespace MoogleEngine
                 document_vectors.Add(new Vector(array));
             }
             Vector query_vector = new(query_array);
-            Dictionary<int, double> coisine_sim = new();
+            Dictionary<int, float> coisine_sim = new();
             for (int i = 0; i < document_vectors.Count; i++)
             {
                 coisine_sim.Add(i, CoisineSimilarity(document_vectors[i], query_vector));
